A CONTACT LIST
This java programing is about a contact list  that you add name and
number and if you want delete or edit them you can. it always deal with
two language what we call in java i18n. the language that are in the
java program are English and French. you can search and change color of
the table.